Detect one more case of auto-unboxing

+++ test/src/com/greentube/convertertest/OverloadedClass2.java
@@ -2,9 +2,9 @@ package com.greentube.convertertest;
 
 public class OverloadedClass2 extends OverloadedClass implements OverloadedInterface2
 {
-    public int plus(int a, int b)
+    public double plus(double a, double b)
     {   
-        return a + b + 2;         
+        return a + b + 2.0;
     }
 
     public static double add(double a, double b)
